+++ internal/ipsec/ipsec_test.go
@@ -28,14 +28,14 @@ func canonicalIntent() *ipsec.TunnelIntent {
 			{ // A = FortiGate, public WAN, dynamic responder (peer B is behind NAT)
 				DeviceID: 1, Vendor: "fortigate", PeerIP: "66.179.9.155",
 				EgressIface: "port1", LANIface: "port3",
-				LocalID:          ipsec.IKEIdentity{Type: ipsec.IDTypeKeyID, Value: "fwm-t7-a"},
+				LocalID:          ipsec.IKEIdentity{Type: ipsec.IDTypeFQDN, Value: "fwm-t7-a"},
 				ProtectedSubnets: []string{"10.10.10.0/24"},
 				InnerIP:          innerA, Reqid: 7, MSSClamp: 1350, ChildLifetimeSecs: 7200,
 			},
 			{ // B = OPNsense, private WAN, sole initiator
 				DeviceID: 2, Vendor: "opnsense", PeerIP: "192.168.5.107", Dynamic: true,
 				EgressIface: "wan", LANIface: "lan",
-				LocalID:          ipsec.IKEIdentity{Type: ipsec.IDTypeKeyID, Value: "fwm-t7-b"},
+				LocalID:          ipsec.IKEIdentity{Type: ipsec.IDTypeFQDN, Value: "fwm-t7-b"},
 				ProtectedSubnets: []string{"192.168.50.0/24"},
 				InnerIP:          innerB, Reqid: 7, MSSClamp: 1350, ChildLifetimeSecs: 3600,
 			},